RPCT-1031, u7: TuningClient data has been changed from double to float

+++ TuningClient/DialogInputValue.h
@@ -12,21 +12,21 @@ class DialogInputValue : public QDialog
 	Q_OBJECT
 
 public:
-    explicit DialogInputValue(bool analog, double value, bool sameValue, double lowLimit, double highLimit, int decimalPlaces, QWidget *parent = 0);
+    explicit DialogInputValue(bool analog, float value, bool sameValue, float lowLimit, float highLimit, int decimalPlaces, QWidget *parent = 0);
 	~DialogInputValue();
 
 private:
 	Ui::DialogInputValue *ui;
 
-	double m_value = 0;
-    double m_lowLimit = 0;
-    double m_highLimit = 0;
+    float m_value = 0;
+    float m_lowLimit = 0;
+    float m_highLimit = 0;
 	bool m_analog = true;
 
 	virtual void accept();
 
 public:
-	double value() { return m_value; }
+    float value() { return m_value; }
 
 private slots:
 	void on_m_checkBox_clicked(bool checked);
